Implements the edit commands
Needs some cleanup on code reuse and extraneous console output.

Ran into trouble trying to figure out how to launch vim on/before
program exit, and lo and behold someone else was too. The answer in this
Stack Overflow question (http://stackoverflow.com/a/21515959/122291)
cleared it up.

You can just execute the command. vim won't run, apparently, if it's not
in a tty, and if it did it'd be completely useless anyhow. So the Stdin
and Stdout values for the Command need to be set to stdin and stdout for
the system and the commands executed there.